+++ src/menu.tcl
@@ -77,16 +77,12 @@ proc menu::SetTint {tint} {
   #   tint - value of the tint
 
   namespace upvar ::alited al al obPav obPav
-#!  namespace upvar ::alited obFND obFND        ;# TODO TLabelFS not tinted immediately
-#!  set ef [winfo exists $alited::find::win]    ;# TODO TLabelFS ...
-#!  if #\{$ef#\} #\{$obFND res $alited::find::win 0#\}  ;# TODO TLabelFS ...
   $obPav csToned $al(INI,CS) $tint yes
   alited::file::MakeThemHighlighted
   alited::main::ShowText
   alited::bar::BAR update
   CheckTint
   alited::ini::initStyles
-#!  if #\{$ef#\} #\{after idle #\{after 10 alited::find::_run#\}#\}  ;# TODO TLabelFS ...
 }
 #_______________________
 
@@ -94,8 +90,9 @@ proc menu::MapRunItems {fname} {
   # Gets a map list to map %f & %D wildcards to the current file & directory names.
   #  fname - the current file name
 
+  namespace upvar ::alited al al
   set ftail [file tail $fname]
-  return [list %D [file dirname $fname] %f $fname %F $ftail \$::FILETAIL $ftail]
+  return [list %PD $al(prjroot) %D [file dirname $fname] %f $fname %F $ftail \$::FILETAIL $ftail]
 }
 #_______________________